fix(jAcl2) regression when removing a user

+++ lib/jelix-modules/jacl2db/classes/jAcl2DbUserGroup.class.php
@@ -267,8 +267,10 @@ class jAcl2DbUserGroup
         // remove from all the groups (jacl2_users_group)
         jDao::get('jacl2db~jacl2usergroup', 'jacl2_profile')->deleteByUser($login);
 
-        // remove the user's personal group (jacl2_group)
-        $daogroup->delete($privategrp->id_aclgrp);
+        if ($privategrp) {
+            // remove the user's personal group (jacl2_group)
+            $daogroup->delete($privategrp->id_aclgrp);
+        }
     }
 
     /**
